move spectator view sharing items into sharing directory
