+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/Sharing.meta
