+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/Sharing/MarkerSpatialCoordinateService.cs
@@ -12,7 +12,7 @@ using Microsoft.MixedReality.Toolkit.Extensions.MarkerDetection;
 using Microsoft.MixedReality.Toolkit.Extensions.Sharing;
 using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Utilities;
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.SpatialCoordinates
+namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Sharing
 {
     public enum MarkerSpatialCoordinateServiceOverlayState
     {
