+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/Sharing/MarkerSpatialCoordinateService.cs.meta
