+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/Sharing/UDPBroadcastNetworkingService.cs
@@ -11,7 +11,7 @@ using UnityEngine;
 using Microsoft.MixedReality.Toolkit.Extensions.Sharing;
 using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Utilities;
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Networking
+namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Sharing
 {
     public delegate void UDPBroadcastConnectHandler(int serverPort, int clientPort);
