--- Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/SpatialCoordinates.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4aa0021c86554234e88510c411eae588
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
